requirements.txt
streamlit==1.28.1
pdfplumber==0.9.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
requests==2.31.0

# Optional for better performance
accelerate==0.24.1
datasets==2.14.6

# For deployment
watchdog==3.0.0